requirements.txt
fastapi
uvicorn
pillow
torch
transformers
opencv-python
python-multipart